pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.0",
    "flask-mail>=0.10.0",
    "fpdf2>=2.8.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
